requirements.txt
pandas
influxdb-client
python-dotenv
requests
zipfile36
